Convert the project to pure ARC
Turn on ARC for the Mac and iOS builds - but not for the tests.

Clear up memory managemnt code, and remove all the finalize methods
because they are a holdover from GC. A lot of the deallocs were cleared
up as well, leaving just not Cocoa cleanup code.

Turn on the Static Analyser.

The iOS library and tests are failing, but this seems to be a long
running problem - as the same problems are present in much earlier
commits.

That shouldn't stop me from pushing on with the desktop code.