Build libgit2 as a static library target of ObjectiveGitFramework
The target builds using a `cmake` shell script stage, being the upstream build system for libgit2.  Building within Xcode as an internal project/target is complicated due to IDE munging of include search paths and duplicated file names in the `src/` and `include/` directories, and so was discounted.

This negates the need to have the binary committed in to the repository, allowing for easier tracking of upstream changes.

The shell script copies the resulting library into the appropriate directory for the project's derived data, so it can still be used as normal with standard Xcode target dependencies, link stages, and build actions (i.e. 'clean').