Remove our duplicated fixture setup.
We seemed to have 2 completely different codepaths, which depending on
which you picked, would lead to a different set of fixtures being used
by your test. This was also the root cause of our faulty clean up.

Now there is but one set of fixtures and it's set up and torn down
properly.