Added note setup step
I had to manually add a header search path as per https://github.com/libgit2/objective-git/issues/441 in order to compile. Added that as a step to the instructions. Only tested on iOS so added to end of that list. Not sure if it also applies for Mac targets (assuming so). If so, please feel free to move to the top where it might make more sense and apply to both.